corrections and refractions

+++ clustering/scripts/plot_metrics.py
@@ -38,34 +38,37 @@ if __name__ == "__main__":
     # reading output files
     path_to_file = pathlib.Path(__file__).parent.parent.absolute()
     output_folder = path_to_file/'output'
-    images_folder = path_to_file/'images'/prefix
+    images_folder = path_to_file/'results'/prefix
     images_folder.mkdir(exist_ok=True)
+    # getting autoencoder results if exist
     clients = sorted(output_folder.glob('*_ae.dat'))
-    # building dataframe of results
-    ae_df = pd.DataFrame()
-    for client in clients:
-        c = pd.read_csv(client)
-        ae_df = ae_df.append(c)
-        client.unlink()
+    ae_df = None
+    if len(clients) > 0:
+        # building dataframe of results
+        ae_df = pd.DataFrame()
+        for client in clients:
+            c = pd.read_csv(client)
+            ae_df = ae_df.append(c)
+            client.unlink()
+        ae_overall = ae_df.copy()
+        ae_overall['client'] = 'all'
+        ae_metrics = list(ae_df.columns)
+        ae_metrics.remove('client')
+        ae_metrics.remove('round')
+    # getting iteration results
     clients = sorted(output_folder.glob('*.dat'))
-    # building dataframe of results
-    df = pd.DataFrame()
-    for client in clients:
-        c = pd.read_csv(client)
-        df = df.append(c)
-        client.unlink()
-        
-    # %%
-    overall = df.copy()
-    overall['client'] = 'all'
-    ae_overall = ae_df.copy()
-    ae_overall['client'] = 'all'
-    metrics = list(df.columns)
-    metrics.remove('client')
-    metrics.remove('round')
-    ae_metrics = list(ae_df.columns)
-    ae_metrics.remove('client')
-    ae_metrics.remove('round')
+    if len(clients) > 0:
+        # building dataframe of results
+        df = pd.DataFrame()
+        for client in clients:
+            c = pd.read_csv(client)
+            df = df.append(c)
+            client.unlink()
+        overall = df.copy()
+        overall['client'] = 'all'
+        metrics = list(df.columns)
+        metrics.remove('client')
+        metrics.remove('round')
     '''
     # %%
     sns.set_style('whitegrid')
@@ -95,20 +98,21 @@ if __name__ == "__main__":
         plt.draw()
         plt.savefig(filename)
         plt.close()
-        
-    tmp = ae_df.copy()
-    tmp = tmp.append(ae_overall)
-    for metric in ae_metrics:
-        title = 'autoencoder '+metric
-        filename = images_folder/str('autoencoder_'+metric)
-        fig, ax = plt.subplots(nrows=1, ncols=1, figsize=(20,10))
-        plt.title(title)
-        ax.set_ylabel(metric)
-        plt.xlabel("round")
-        sns.lineplot(x='round', y=metric, hue='client', data=tmp)#, palette=['Blue', 'Red'])#, style='client')#, markers=['.', '.'])
-        plt.draw()
-        plt.savefig(filename)
-        plt.close()
+    
+    if ae_df is not None:
+        tmp = ae_df.copy()
+        tmp = tmp.append(ae_overall)
+        for metric in ae_metrics:
+            title = 'autoencoder '+metric
+            filename = images_folder/str('autoencoder_'+metric)
+            fig, ax = plt.subplots(nrows=1, ncols=1, figsize=(20,10))
+            plt.title(title)
+            ax.set_ylabel(metric)
+            plt.xlabel("round")
+            sns.lineplot(x='round', y=metric, hue='client', data=tmp)#, palette=['Blue', 'Red'])#, style='client')#, markers=['.', '.'])
+            plt.draw()
+            plt.savefig(filename)
+            plt.close()
     '''
     # %%
     for metric in metrics:
